가독성, 좌석검증 조건 추가에 유연하게 대처, 인자 검증을 초기에 추가함으로써 불필요한 연산 방지를 목적으로 리팩토링 진행

+++ src/main/java/com/example/demo/domain/Screening.java
@@ -14,10 +14,10 @@ import java.time.LocalDateTime;
 public class Screening {
     @Id
     @GeneratedValue
-    private Long screeningId;
+    private Integer screeningId;
 
     @Column(name = "movieId")
-    private Long movieId;
+    private Integer movieId;
     @Column(name = "startTime")
     private String startTime;
     @Column(name = "endTime")
