+++ src/main/java/com/example/demo/service/ScreeningService.java
@@ -20,17 +20,17 @@ public class ScreeningService {
     @Qualifier("screeningJpaRepository")
     private final ScreeningRepository screeningJpaRepository;
 
-    public Screening getScreeningById(Long screeningId){
+    public Screening getScreeningById(Integer screeningId){
         return screeningJpaRepository.findById(screeningId)
                 .orElseThrow(() -> new EntityNotFoundException(String.format("해당 시간대에 영화 없음")));
     }
 
-    public void assertScreeningExists(Long screeningId){
+    public void assertScreeningExists(Integer screeningId){
         if(screeningJpaRepository.existsById(screeningId))
             throw new EntityNotFoundException(String.format("해당 시간대에 영화 없음"));
     }
 
-    public List<Screening> getScreeningListByMovieId(Long movieId){
+    public List<Screening> getScreeningListByMovieId(Integer movieId){
         return screeningJpaRepository.findAllByMovieId(movieId)
                 .orElseThrow(() -> new EntityNotFoundException(String.format("상영 정보 없음")));
     }
